Move private retry helpers into agent/__init__.py to silence reportPrivateUsage
